Started reorganization for work on the reactor version of lnsd.
Basing lnsd off of the reactor is meant to make it run in a single thread
(versus having to do manual locking), as well as to make it portable to
Windows. I'm dropping the SOCKS server, since I don't use it, and it
could be rewritten to be external to lnsd using lnsd's socket protocol.