lnsd is now capable of handling hosts with the same name.
In order to handle this the JSON protocol, between the control handler
on the server and the client, was changed to allow IP messages to
contain lists of addresses rather than just one address. The message
which retrieves the entire mapping now sends back hosts mapped to lists
of IP addresses, rather than hosts mapped to a single address.