Changed the configuration options that lnsd and lns-query take.
I decided to remove the timeout and heartbeat options, since 10s and 30s
are sensible defaults, and because cases where machines disagree can
possibly cause hosts to drop while they still exist on the network.

For example, a host can have a timeout of 10s while another has a
heartbeat timer of 12s, and the host with the 10s timeout will see the
host with the 12s heartbeat connect and disconnect from the network
repeatedly. Standardizing the heartbeat and timeout parameters prevent
this situation from occurring.

Also, both the external network port can be set, as well as the internal
port used for communication between the client and the server.
Previously, only the external port could be set.